+++ src/Command/Test/Dockerfiles.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace App\Command\Test;
 
+use App\Command\Dockerize;
+use Symfony\Component\Console\Input\InputOption;
+
 class Dockerfiles extends AbstractMultithreadTest
 {
     /**
@@ -13,7 +16,12 @@ class Dockerfiles extends AbstractMultithreadTest
     {
         $this->setName('test:dockerfiles')
             ->setDescription('<info>Install Magento with local Dockerfiles to test them after changes are made</info>')
-            ->setHelp(<<<'EOF'
+            ->addOption(
+                Dockerize::OPTION_EXECUTION_ENVIRONMENT,
+                'e',
+                InputOption::VALUE_REQUIRED,
+                'Use local Dockerfile from the Docker Infrastructure repository instead of the prebuild DockerHub image'
+            )->setHelp(<<<'EOF'
 The <info>%command.name%</info> sets up multiple Magento versions with the defined local Dockerfiles to test them.
 Includes installing Sample Data and reindex.
 
@@ -32,28 +40,34 @@ EOF);
 
     /**
      * @param string $domain
-     * @param string $phpVersion
      * @param string $magentoVersion
-     * @throws \JsonException
+     * @param string $phpVersion
      */
-    protected function installMagento(string $domain, string $phpVersion, string $magentoVersion): void
+    protected function installMagento(string $domain, string $magentoVersion, string $phpVersion): void
     {
-        throw new \RuntimeException('In development');
+        $projectRoot = $this->env->getProjectsRootDir() . $domain;
+        $executionEnvironment = $this->input->getOption(Dockerize::OPTION_EXECUTION_ENVIRONMENT);
+
         $this->shell->exec(<<<BASH
             php {$this->getDockerizerPath()} magento:setup $magentoVersion \
-                --domains="$domain www.$domain" --php=$phpVersion -nf
+                --domains="$domain www.$domain" \
+                --php=$phpVersion \
+                --execution-environment=$executionEnvironment \
+                -nf
         BASH);
         $this->shell->exec(
             <<<BASH
-                git add .gitignore .htaccess docker* var/log/ app/
-                git commit -m "Docker and Magento files after installation" -q
+                git add -A
+                git commit -m "Initial commit (with Docker infrastructure)" -q
             BASH,
-            $projectRoot
+            $projectRoot,
+            true
         );
 
-        $this->shell->dockerExec('php bin/magento sampledata:deploy', $domain);
-        $this->shell->dockerExec('php bin/magento setup:upgrade', $domain);
-        $this->shell->dockerExec('php bin/magento deploy:mode:set production', $domain);
+        $this->execWithTimer("docker exec $domain php bin/magento sampledata:deploy 2>/dev/null", true);
+        $this->execWithTimer("docker exec $domain php bin/magento setup:upgrade");
+        $this->execWithTimer("docker exec $domain php bin/magento deploy:mode:set developer");
+        $this->execWithTimer("docker exec $domain php bin/magento indexer:reindex");
 
         $this->log("Website address: https://$domain");
     }