+++ Changelog.md
@@ -11,13 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New `test:dockerfiles` command to test running different Magento versions before publishing the Dockerfiles.
-- Implemented `--dockerfile` (`-d`) option for `magento:setup` and `dockerize` commands. Must be used only for testing! Use prebuild images for yor projects;
+- Implemented `--execution-environment` (`-e`) option for `magento:setup` and `dockerize` commands. Must be used only for testing! Use prebuild images for yor projects;
 
 ### Changed
 
 - Renamed command from `hardware:test` to `test:hardware` and mover the class to `App\Command\Test` namespace.
 - Extracted all common functionality for the `test:hardware` command into an abstract class (compatibility-breaking change).
 - Moved all logs to the same location - `var/log/`.
+- Git user name and email, Magento admin user name and email changed to the neutral ones.
 
 
 ## [2.0.0] - 2020-05-21